Error when using relativeCaption plugin
This issue comes from lightgallery itself.
Definition of LightGalleryAllSettings is missing RelativeCaptionSettings.

A pull-request has been proposed as
https://github.com/sachinchoolur/lightGallery/pull/1724

In the meantime, a tempory fix is also proposed in astro-lightgallery.
It consists in defining a new LightGalleryAllSettings, called
LightGalleryAllSettingsFix, that includes LightGalleryAllSettings
and the settings of the plugin Relative Caption and Vimeo Video.

Once the pull-request in lightgallery is accepted, this one
must be removed.

Fixes #1
Signed-off-by: Pascal Brand <pascal.brand38@gmail.com>